+++ tests/jasmine/behavior/generated-parser-behavior.spec.js
@@ -221,6 +221,39 @@ describe("generated parser behavior", function() {
       });
     });
 
+    describe("non-start rule", function() {
+
+      describe("when the expression doesn't match", function() {
+        describe("without display name", function() {
+          it("reports match failure and records an expectation", function() {
+            var parser = PEG.buildParser(
+              'start = alpha "b";\n' +
+              'alpha = a+;\n' +
+              'a = "a";\n'
+            );
+
+            expect(parser).toFailToParse("b", {
+              expected: [{ type: "literal", value: "a", description: '"a"' }]
+            });
+          });
+        });
+
+        describe("with display name", function() {
+          it("silences any expectations recorded when matching the expression", function() {
+            var parser = PEG.buildParser(
+              'start = alpha "b";\n' +
+              'alpha "alpha" = a+;\n' +
+              'a = "a";\n'
+            );
+
+            expect(parser).toFailToParse("c", {
+              expected: [{ type : 'other', description : 'alpha' }]
+            });
+          });
+        });
+      });
+    });
+
     describe("positive semantic predicate", function() {
       describe("initializer variables & functions", function() {
         it("can access variables defined in the initializer", function() {